[#14] Remove length calculation
Clojurescript doesn't have a method to compute the size of protobuf messages, and
we can get away without having one by using auto-expanding ByteArrayOutputStream
types.  We therefore remove support for all the size-XX serdes and (pb/length)
functions.  This is part of the breaking changes to support both Clojure
and Clojurescript

Partially addresses #14

Signed-off-by: Greg Haskins <greg@manetu.com>