Add 'close-ch' feature
Streaming GRPC servers need to be able to handle the case where their
client has disconnected and is no longer listening. Other languages
provide a language appropriate signal when this has occured, but we currently
lack a way to do this with Protojure.

This patch adds a new feature where a core.async promise-chan is added
to the request map as :close-ch.  The protocol for this promise-chan
is quite simple.  It will remain unresolved while the client is still
connected, and resolve to "true" if they disconnect.

This feature is generally only needed for streaming servers
since the app-handler is probably in some kind of go-loop, waiting
for items to stream to the client.  This feature means that the
app-handler now has the option to alt on the close-ch so that they
may be informed of, and react to, a client terminating the connection
out from under them.

Signed-off-by: Greg Haskins <greg@manetu.com>