Add support for connection-level metadata
This patch adds a "metadata" option to grpc.http2 client
objects, helpful for passing http headers such as openid
connect authorization tokens.

This feature was exposed as a per-invoke option, but the
protoc-gen-clojure compiler doesnt currently surface it
up through the generated code.  This patch may be used
as a work-around, at least for connection-level granularity.

The biggest downside is that if a token expires, the client
will need to connect again to refresh it.  We may enhance this
in the future.

Signed-off-by: Greg Haskins <greg@manetu.com>