Ensure we close the output channel under error conditions
This fixes a bug where a very early error in a grpc (invoke)
may leave the output channel in an open state, which can lead
to hung clients.  The fix is to ensure that we close the
channel in the error path.  The happy path closes the channel
via (receive-payload).

N.B. It may be smarter to close the channel in a (finally) block
after the exception regardless of success or failure, but further
analysis is needed to ensure this is indeed safe.

N.B. The new channel-close logic seems to have regressed coverage
slightly.  I believe its how cloverage rates the timbre logging
as yellow.  Since the new logic is covered by UTs, I'm not particuarly
concerned.

Signed-off-by: Greg Haskins <greg@manetu.com>