Implement proper connection-close subscription in Undertow
The current logic has a problem in that it registers a per exchange
subscription to a per-connection resource.  This can result in effectively
a resource leak if a client employs a 1:N connection:request model, which
http2 encourages.

This patch addresses this by building an explicit per-exchange subscription
substrate, thereby allowing resources to be freed based on the lifecycle
of a request idependently from the connection lifecycle.

Signed-off-by: Greg Haskins <greg@manetu.com>