Fix leak in unsubscribe
The previous fix only partially addressed the per-exchange leak that
we were observing.  This should fix that logic such that the request
properly unsubscribes on completion.

There is still potentially another issue that remains related to an
obervation of an unexpectedly high number of registered connections,
which may mean that we are not properly detecting when a connection
is already listening, TBD.

Signed-off-by: Greg Haskins <greg@manetu.com>