Remove elision for empty messages
According to the proto3 spec, {} is semantically different from nil. On
top of potentially being relevant in edge cases, this is important
because a `nil` message in a oneof means the entire oneof is nil, while
an empty message still sets the corresponding oneof, meaning that empty
messages can (and sometimes are) be used as sentinel values in oneofs.

Signed-off-by: Sean Harrap <sean.harrap@gmail.com>