Fix flow-control issue with later versions of Jetty
It's not entirely clear why, but older versions of jetty handled flow-control
differently w.r.t. stream vs session windows.  In later versions, we were getting
a RST/GOAWAY sequence once the session window was depleted without sending a session-level
WINDOW_UPDATE, terminating the connection.

This patch restores the ability to have protojure clients work with newer Jetty.  It's not
entirely clear what Jetty will do when the default session-window depletes, but we are now
riding on defaults which is hopefully the highest probability that the correct thing
happens.  More investigation needed to confirm this.

Signed-off-by: Greg Haskins <greg@manetu.com>