Properly capture exceptions in the pedestal stack
(p/resolved) will not trap exceptions and re-cast them
to p/catch, but we were assuming it would.  This patch
converts those callsites to a full p/promise which does
properly forward exceptions.  The net result is that a
client-related crash, such as a CloseChannelException
is now handled in the manner that we intended, which is
to say that its graceful.

Signed-off-by: Greg Haskins <greg@manetu.com>