Update to the latest promesa (#102)
N.B. While the protojure API is not changing, there are significant
changes to promesa between v3.0.0 and v6.0.2 that may become visible
to your application if you were also consuming promesa.

These changes were largely introduced in the v4.0.0 release of promesa,
which you can find more details about here:

https://cljdoc.org/d/funcool/promesa/6.0.2/doc/changelog#version-400

The most apparent change is the switch from p/promise to p/create.  However
a more subtle change is related to the default threading model.  v3.x
implicitly ran p/then and other chaining methods in a forkjoin pool.  Starting
in v4.x and above, the executor is now configurable and the default is
caller-thread based.  This should result in higher performance but it may
break existing applications if they happen to be exploiting the async
nature of the forkjoin pool, even if inadvertently.

Signed-off-by: Greg Haskins <greg@manetu.com>