Re-write LPM IO layer to use ByteBuffer orientation
The current code uses a byte-wise strategy for the LPM serdes
transport.  It was known going in that this was probably
suboptimal, but we were going for correct first, fast
second.

Profiling has demonstrated that this is indeed a hotspot,
which a lot of overhead being spent in writing single
bytes into core.async channels.

This patch refactors the LPM layer to use a ByteBuffer oriented
approach, which significantly reduces the overhead for moving
bytes around.  We make gains because of several net benefits:

1) Orders of magnitude fewer interactions with core.async put!
2) Opens up byte[]-wise operators in the XXStream interfaces
3) Fewer memory copies as we can ByteBuffer/flip buffers around

We also retune some of the parameters for HTTP2 buffer/flow-control.

Results from bidi test are

"Elapsed time: 231.706687 msecs"

which is approximately a 15x improvement in performance

Signed-off-by: Greg Haskins <greg@manetu.com>