Split into [io core grpc-client grpc-server] modules
This allows us to solve a few problems:

1. IO streams require AOT but now we can limit AOT to just the protojure/io lib
2. Use cases involving only GRPC clients do not need to bring in undertow/pedestal
3. Use cases involving only GRPC servers do not need to bring in jetty-client

N.B. We are losing two features of the previous build

A. Cloverage: its not clear how to get cloverage to give decent coverage reports in a multi-module
   project without refactoring and localizing the tests.  We leave that as a future task and drop
   back to plain eftest without coverage reporting.
B. We disable codox, which was not working properly anyway.  We can add it back in the future.

Signed-off-by: Greg Haskins <greg@manetu.com>