[io] Eliminate AOT with native .java classes/interfaces
I thought the split would be enough to solve the AOT w/gen-class issue, but after
further investigation, it turns out that it is not becaues we were still bundling
clojure and core.async .class files.  This patch eliminates the problem entirely
by taking a different approach with native java.  This was found to be neessary
because of the weirdness of java.io's use of multi-arity overloading that precluded
non gen-class solutions such as clojure.core/proxy.

Signed-off-by: Greg Haskins <greg@manetu.com>