disconnect: Return a promise that blocks until the connection closes
The `.stop` method on the Jetty HTTP2Client is non-blocking. The client
must use an event listener to synchronize the state until the client
completely shuts down.

This changes the interface of the public `(disconnect)` API to return
a promise instead, which will resolve when the HTTP2Client finishes
stopping (or throws on an error).

Signed-off-by: Ryan Sundberg <ryan@arctype.co>