Automatically determine ssl parameter when uri has https scheme (#134)
Convenience enhancement so caller doesn't have to specify :uri for common
situations.  If uri starts with https:// scheme, then use ssl as true.
Still respect provided :ssl param if present.

Added tests for this situation that use circleci.bond for this purpose,
since it's difficult to setup an SSL test; this verifies the expected
connect call is made.

Signed-off-by: James Vickers <james.vickers@reifyhealth.com>