Fix handler threading
This patch fixes a bug where the response logic was inadvertently running within core.async thread
context.  I believe this was a side-effect of how promesa schedules p/then in conjunction with the
go-routine in transmit-trailers...the finally clause was piggy-backing on the core.async thread
when we intended it to run from the thread-pool.

This patch ensures that we return to the thread-pool context before executing the finally clause.

The fix has been confirmed via yourkit profiling.

Signed-off-by: Greg Haskins <greg@manetu.com>