Use case-insensitive regex instead of bulk-converting all headers to lower-case
We added logic that converted all headers to lower-case to make finding certain
headers, such as content-type and grpc-encoding, work in a case-insensitive manner.

However, this approach causes problems with case-sensitive headers like the X-B3
trace propagators.  This change removes the bulk conversion to lower-case in favor
of broader case-insensitive pattern matching.

N.B.  This patch may be a breaking change if any applications built on
Protojure's grpc-server relied on the header behavior.  As such,
we will bump the version for this patch.

Signed-off-by: Greg Haskins <greg@manetu.com>